+++ tests/version-validator.sh
@@ -86,6 +86,9 @@ expect_pass '2.0.0-rc13.1'
 write_fixture '2.0.0-rc13.2' '2.0.0-rc13.2' '1.0.0'
 expect_pass '2.0.0-rc13.2'
 
+write_fixture '2.0.0-rc13.3' '2.0.0-rc13.3' '1.0.0'
+expect_pass '2.0.0-rc13.3'
+
 write_fixture '2.0.0-rc1' '2.0.0-rc1' '2.0.0'
 expect_fail '2.0.0-rc1'
 
@@ -124,6 +127,9 @@ expect_fail '2.0.0-rc9'
 
 write_fixture '2.0.0-rc13.1' '2.0.0-rc13.2' '1.0.0'
 expect_fail '2.0.0-rc13.1'
+
+write_fixture '2.0.0-rc13.2' '2.0.0-rc13.3' '1.0.0'
+expect_fail '2.0.0-rc13.2'
 
 printf 'Release version validator tests passed.\n'